Divide Payment Page into Components
