+++ frontend/src/styles/home/items.css
@@ -2,7 +2,7 @@
   height: 100%;
   width: 100%;
   overflow: hidden;
-  background-image: url("../img/pixels.jpg");
+  background-image: url("../../img/pixels.jpg");
   background-size: cover;
   padding: 0 !important;
