+++ frontend/src/styles/home/loaderDonators.css
