+++ frontend/src/styles/home/loaderItems.css
