+++ frontend/src/styles/home/slider.css
