+++ frontend/src/styles/payment/buttonHeader.css
@@ -1,0 +1,62 @@
+.button-arounder {
+  width: 483px;
+  font-size: 32px;
+  background: hsl(190deg, 30%, 15%);
+  color: hsl(190deg, 10%, 95%);
+  border: solid 2px black;
+  padding: 0.375em 1.125em;
+  margin: 15px;
+  box-shadow: 0 0px 0px hsla(190deg, 15%, 5%, 0.2);
+  transform: translateY(0);
+  border-top-left-radius: 0px;
+  border-top-right-radius: 0px;
+  border-bottom-left-radius: 0px;
+  border-bottom-right-radius: 0px;
+
+  --dur: 0.15s;
+  --delay: 0.15s;
+  --radius: 16px;
+
+  transition: border-top-left-radius var(--dur) var(--delay) ease-out,
+    border-top-right-radius var(--dur) calc(var(--delay) * 2) ease-out,
+    border-bottom-right-radius var(--dur) calc(var(--delay) * 3) ease-out,
+    border-bottom-left-radius var(--dur) calc(var(--delay) * 4) ease-out,
+    box-shadow calc(var(--dur) * 4) ease-out,
+    transform calc(var(--dur) * 4) ease-out,
+    background calc(var(--dur) * 4) steps(4, jump-end);
+}
+
+.button-arounder:hover,
+.button-arounder:focus {
+  box-shadow: 0 4px 8px hsla(190deg, 15%, 5%, 0.2);
+  transform: translateY(-4px);
+  background: hsl(230deg, 50%, 45%);
+  border-top-left-radius: var(--radius);
+  border-top-right-radius: var(--radius);
+  border-bottom-left-radius: var(--radius);
+  border-bottom-right-radius: var(--radius);
+}
+
+@media (min-width: 1024px) and (max-width: 1100px) {
+  .button-arounder {
+    width: 340px;
+    margin: 6px;
+    font-size: 28px;
+  }
+}
+@media (min-width: 1100px) and (max-width: 1200px) {
+  .button-arounder {
+    width: 416px;
+  }
+}
+@media (min-width: 1200px) and (max-width: 1300px) {
+  .button-arounder {
+    width: 460px;
+  }
+}
+@media (min-width: 720px) and (max-height: 540px) {
+  .button-arounder {
+    width: 259px;
+    font-size: 21px;
+  }
+}
