+++ frontend/src/styles/payment/donateInfo.css
@@ -1,0 +1,135 @@
+.donate-card {
+  height: 100%;
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  font-family: "Poppins", sans-serif;
+  background-color: #5b5b5b;
+
+  .image {
+    overflow: hidden;
+    img {
+      width: 100%;
+      transform-origin: center;
+      transform: scale(var(--img-scale));
+      transition: transform 0.4s ease-in-out;
+    }
+  }
+  .donate-desc {
+    width: 90%;
+    padding: 15px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    color: var(--title-color);
+
+    h2 {
+      font-family: "Courier New", Courier, monospace;
+      color: var(--title-color);
+      margin: 0 0 18px 0;
+      font-size: 1.9rem;
+      letter-spacing: 0.06em;
+      transition: color 0.3s ease-out;
+    }
+    .text-left,
+    .text-right {
+      width: 100%;
+      font-size: 18px;
+      font-weight: 400;
+      display: flex;
+      justify-content: space-between;
+      a {
+        color: var(--title-color);
+        text-decoration: none;
+      }
+      a::after {
+        position: absolute;
+        inset-block: 0;
+        inset-inline: 0;
+        cursor: pointer;
+        content: "";
+      }
+    }
+
+    .text-right {
+      justify-content: flex-end;
+      font-size: 20px;
+      font-weight: 600;
+    }
+    .progress-desc {
+      width: 100%;
+      padding-top: 20px;
+
+      .progress {
+        background: rgba(255, 255, 255, 0.1);
+        justify-content: flex-start;
+        border-radius: 50px;
+        align-items: center;
+        position: relative;
+        padding: 0 3px;
+        display: flex;
+        height: 20px;
+        width: 100%;
+      }
+
+      .progress-value {
+        width: 0; /* Initial width will be set dynamically */
+        transition: width 0.8s ease-in-out; /* Smooth transition */
+        box-shadow: 0 10px 40px -10px #fff;
+        border-radius: 50px;
+        background: var(--title-color);
+        height: 20px;
+      }
+
+      .progress-text {
+        width: 100%;
+
+        .donated-text,
+        .goal-text {
+          color: var(--title-color);
+        }
+
+        .donated-text {
+          float: left;
+          font-size: 20px;
+          font-weight: 600;
+          font-family: "Times New Roman", Times, serif;
+        }
+
+        .goal-text {
+          float: right;
+          font-size: 16px;
+          font-family: "Times New Roman", Times, serif;
+        }
+      }
+    }
+  }
+}
+
+@media (min-width: 720px) and (max-height: 540px) {
+  .donate-card .donate-desc {
+    h2 {
+      font-size: 20px;
+    }
+    .text-left,
+    .text-right {
+      font-size: 11px;
+    }
+    .text-right {
+      font-size: 11px;
+    }
+    .progress-desc {
+      padding-top: 15px;
+
+      .progress-text {
+        .donated-text {
+          font-size: 12px;
+        }
+        .goal-text {
+          font-size: 12px;
+        }
+      }
+    }
+  }
+}
