+++ frontend/src/styles/payment/monoMethod.css
